연관관계 설정 - (파킹노드 엔티티-워크인포엔티티)

+++ backend/src/main/java/com/iot2ndproject/mobilityhub/domain/parkingmap/repository/ParkingMapEdgeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ParkingMapEdgeRepository extends JpaRepository<ParkingMapEdgeEntity, Long> {
+public interface ParkingMapEdgeRepository extends JpaRepository<ParkingMapEdgeEntity, Integer> {
 
     List<ParkingMapEdgeEntity> findByFromNode(ParkingMapNodeEntity fromNode);
 }
